[#38] move pause-button to shared ui
