+++ src/shared/ui/pause-button/index.ts
@@ -1,1 +1,2 @@
+// TODO: need to move component to shared/ui
 export { default as PauseButton } from './pause-button.vue'
